Understanding the ABCD of Web Development
I started with opening the account on GitHub and downloaded the desktop version of the same.
After that, I started coding on Visual Studio.
At first, I struggled a little to create a solid project but then decided to open two separate files.
1. html
2.css

1. On the assessment html file, I used the codes I discovered and those available in visual studio.
A) At first, I filled in the the 'Title'- which appears on the page when I open. Then, I put heading with <h1> inside <body>
B) Then, I decided to play a little with a different code and put buttons with text in the code.
C) I had to try different codes to finally get at the right one.

2.
On the Assessment CSS file - I coded the colors I wanted to use on the heading.
I also played a little with the font styles . Now, the real struggle was to put the buttons.
For the same, I had to explore codes for the buttons and then tried different styles and colors.

How to Merge html and css in visual studio:
I went to html file and below the title, I coded <link href = "copied path of the css file" rel = "stylesheet" type = "text/css"
This code merged both the files and gave me the result I wanted with structure and style on the webpage.