Learning the basics of collaboration when it comes to web development
I started with opening the account on GitHub and downloaded the desktop version of the same.
After that, I started coding on Visual Studio.
At first, I struggled a little to create a solid project but then decided to open two separate files.
1. html
2.css

1. On the assessment html file, I used the codes I discovered and those available in visual studio.
A) At first, I filled in the the 'Title'- which appears on the page when I open. Then, I put heading with <h1> inside <body>
B) Then, I decided to play a little with a different code and put buttons with text in the code.
C) I had to try different codes to finally get at the right one.