Experience with Visual Studio Code
I used Visual Studio initially but it didn't work much for me. Also, the folders I created had some errors.
So, I downloaded Visual Studio Code and fortunately I could save html file, so could work on that.

So, here I tried to add one more button along with visual and auditory, i.e. emotional to refine more on the subject matter.

I also chose to add heading 2.
I wanted to use heading 2 to elaborate more on heading one, which was just and outline of what one can expect.

This is a webpage for information and learning, so developing with larger text size makes it easier to read for a user.

I was searching for a right image to put