Image for the subject
To explain the complex nature of human mind, I have selected galaxy as an image.
I also wanted it to be on larger size, so changed it as well in HTML.
The larger size of galaxy may make the user feel the depth of science involved in understanding a human mind, twirls of emotions and complexity.