Update Assessment1.html
It is important to add a link to validate and make the informative web page more authentic. It adds integrity and reliability.
So, I used an HTML code with text to add a hyperlink. From the entire code, only the text part is going to be visible for the users. Once they click on the text, it will lead them to the webpage.